Move Type parsers to own package 2
